server/requirements.txt
signalrcore>=0.9.5
python-dotenv>=1.0.0
